# openapi.yaml
openapi: 3.0.0
paths:
...
  /products:
    get:
      operationId: listProducts
      parameters:
        - name: page
          required: false
          in: query
          description: Page number (1-based)
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          description: Number of items per page
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: search
          required: false
          in: query
          description: Search term for name or SKU
          schema:
            type: string
        - name: category_id
          required: false
          in: query
          description: Filter by category ID
          schema:
            format: uuid
            type: string
        - name: brand_id
          required: false
          in: query
          description: Filter by brand ID
          schema:
            format: uuid
            type: string
        - name: primary_supplier_id
          required: false
          in: query
          description: Filter by primary supplier ID
          schema:
            format: uuid
            type: string
        - name: is_active
          required: false
          in: query
          description: Filter by active status
          schema:
            type: boolean
        - name: is_perishable
          required: false
          in: query
          description: Filter by perishable status
          schema:
            type: boolean
        - name: min_price
          required: false
          in: query
          description: Minimum price filter
          schema:
            minimum: 0
            type: number
        - name: max_price
          required: false
          in: query
          description: Maximum price filter
          schema:
            minimum: 0
            type: number
        - name: include_deleted
          required: false
          in: query
          description: Include soft-deleted products
          schema:
            default: false
            type: boolean
        - name: sort_by
          required: false
          in: query
          description: Field to sort by
          schema:
            default: name
            type: string
            enum:
              - name
              - sku
              - created_at
              - updated_at
              - standard_price
              - standard_cost
              - reorder_point
        - name: sort_order
          required: false
          in: query
          description: Sort order
          schema:
            default: ASC
            type: string
            enum:
              - ASC
              - DESC
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProductsResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: List products with pagination and filtering
      tags:
        - Products
    post:
      operationId: createProduct
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponseDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Create product
      tags:
        - Products
  /products/all:
    get:
      operationId: listAllProducts
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: List all products without pagination
      tags:
        - Products
  /products/{id}:
    get:
      operationId: getProduct
      parameters:
        - name: id
          required: true
          in: path
          description: Product UUID
          schema:
            type: string
        - name: include_deleted
          required: false
          in: query
          description: Include soft-deleted products
          schema:
            type: boolean
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponseDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Get product by ID
      tags:
        - Products
    put:
      operationId: updateProduct
      parameters:
        - name: id
          required: true
          in: path
          description: Product UUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponseDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Update product
      tags:
        - Products
    delete:
      operationId: deleteProduct
      parameters:
        - name: id
          required: true
          in: path
          description: Product UUID
          schema:
            type: string
        - name: permanent
          required: false
          in: query
          description: Permanently delete instead of soft delete
          schema:
            type: boolean
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponseDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Delete product
      tags:
        - Products
  /products/category/{categoryId}:
    get:
      operationId: getProductsByCategory
      parameters:
        - name: categoryId
          required: true
          in: path
          description: Category UUID
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponseDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Get products by category
      tags:
        - Products
  /products/category/{categoryId}/tree:
    get:
      operationId: getProductsByCategoryTree
      parameters:
        - name: categoryId
          required: true
          in: path
          description: Category UUID
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponseDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Get products by category tree
      tags:
        - Products
  /products/bulk:
    post:
      operationId: bulkCreateProducts
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkCreateProductsDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkOperationResultDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Bulk create products
      tags:
        - Products
    delete:
      operationId: bulkDeleteProducts
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkDeleteDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkOperationResultDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Bulk delete products
      tags:
        - Products
  /products/bulk/status:
    patch:
      operationId: bulkUpdateProductStatus
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkUpdateStatusDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkOperationResultDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Bulk update product status
      tags:
        - Products
  /products/{id}/restore:
    patch:
      operationId: restoreProduct
      parameters:
        - name: id
          required: true
          in: path
          description: Product UUID
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponseDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Restore deleted product
      tags:
        - Products
  /products/bulk/restore:
    patch:
      operationId: bulkRestoreProducts
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkRestoreDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkOperationResultDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Bulk restore products
      tags:
        - Products
...
  /inventory/product/{productId}:
    get:
      operationId: getInventoryByProduct
      parameters:
        - name: productId
          required: true
          in: path
          description: Product UUID
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryResponseDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Get inventory by product
      tags:
        - Inventory
...
info:
  title: LibreStock Inventory API
  description: REST API for LibreStock Inventory Management System
  version: 1.0.0
  contact:
    name: API Support
    url: ''
    email: ''
  license:
    name: MIT
    url: ''
...
servers:
  - url: http://localhost:8080
    description: Development Server
  - url: https://api.librestock-inventory.com
    description: Production Server
components:
  securitySchemes:
    BearerAuth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Clerk JWT token
  schemas:
...
    ErrorResponseDto:
      type: object
      properties:
        error:
          type: string
          example: Resource not found
      required:
        - error
...
    MessageResponseDto:
      type: object
      properties:
        message:
          type: string
          example: Operation completed successfully
      required:
        - message
...
    ProductResponseDto:
      type: object
      properties:
        created_at:
          format: date-time
          type: string
          description: Creation timestamp
        updated_at:
          format: date-time
          type: string
          description: Last update timestamp
        deleted_at:
          type: object
          description: Deletion timestamp (for soft-deleted records)
          format: date-time
          nullable: true
        created_by:
          type: object
          description: User ID who created the record
          nullable: true
        updated_by:
          type: object
          description: User ID who last updated the record
          nullable: true
        deleted_by:
          type: object
          description: User ID who deleted the record
          nullable: true
        id:
          type: string
          description: Unique identifier
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        sku:
          type: string
          description: Product SKU
          example: PROD-001
        name:
          type: string
          description: Product name
          example: Marine Pump
        description:
          type: object
          description: Product description
          nullable: true
        category_id:
          type: string
          description: Category ID
          format: uuid
        category:
          description: Nested category information
          nullable: true
          allOf:
            - $ref: '#/components/schemas/CategorySummaryDto'
        brand_id:
          type: object
          description: Brand ID
          format: uuid
          nullable: true
        volume_ml:
          type: object
          description: Volume in milliliters
          nullable: true
        weight_kg:
          type: object
          description: Weight in kilograms
          nullable: true
        dimensions_cm:
          type: object
          description: 'Dimensions in cm (format: LxWxH)'
          nullable: true
          example: 10x20x5
        standard_cost:
          type: object
          description: Standard cost
          nullable: true
        standard_price:
          type: object
          description: Standard price
          nullable: true
        markup_percentage:
          type: object
          description: Markup percentage
          nullable: true
        reorder_point:
          type: number
          description: Reorder point threshold
          example: 10
        primary_supplier_id:
          type: object
          description: Primary supplier ID
          format: uuid
          nullable: true
        primary_supplier:
          description: Nested supplier information
          nullable: true
          allOf:
            - $ref: '#/components/schemas/SupplierSummaryDto'
        supplier_sku:
          type: object
          description: Supplier SKU
          nullable: true
        barcode:
          type: object
          description: Product barcode
          nullable: true
          example: '0641628607549'
        unit:
          type: object
          description: Unit of measure
          nullable: true
          example: units
        is_active:
          type: boolean
          description: Whether the product is active
        is_perishable:
          type: boolean
          description: Whether the product is perishable
        notes:
          type: object
          description: Additional notes
          nullable: true
        _links:
          description: HATEOAS links
          allOf:
            - $ref: '#/components/schemas/ProductLinksDto'
      required:
        - created_at
        - updated_at
        - id
        - sku
        - name
        - description
        - category_id
        - brand_id
        - volume_ml
        - weight_kg
        - dimensions_cm
        - standard_cost
        - standard_price
        - markup_percentage
        - reorder_point
        - primary_supplier_id
        - supplier_sku
        - barcode
        - unit
        - is_active
        - is_perishable
        - notes
...
    PaginatedProductsResponseDto:
      type: object
      properties:
        data:
          description: List of products
          type: array
          items:
            $ref: '#/components/schemas/ProductResponseDto'
        meta:
          description: Pagination metadata
          allOf:
            - $ref: '#/components/schemas/PaginationMeta'
      required:
        - data
        - meta
    CreateProductDto:
      type: object
      properties:
        sku:
          type: string
          description: Product SKU
          example: PROD-001
          minLength: 1
          maxLength: 50
        name:
          type: string
          description: Product name
          example: Marine Pump
          minLength: 1
          maxLength: 200
        description:
          type: object
          description: Product description
          maxLength: 1000
          nullable: true
        category_id:
          type: string
          description: Category ID
          format: uuid
        brand_id:
          type: object
          description: Brand ID
          format: uuid
          nullable: true
        volume_ml:
          type: object
          description: Volume in milliliters
          nullable: true
          minimum: 1
        weight_kg:
          type: object
          description: Weight in kilograms
          nullable: true
          minimum: 0
        dimensions_cm:
          type: object
          description: 'Dimensions in cm (format: LxWxH, e.g., 10x20x5)'
          maxLength: 50
          nullable: true
          example: 10x20x5
        standard_cost:
          type: object
          description: Standard cost
          nullable: true
          minimum: 0
        standard_price:
          type: object
          description: Standard price (must be >= standard_cost)
          nullable: true
          minimum: 0
        markup_percentage:
          type: object
          description: Markup percentage
          nullable: true
          minimum: 0
          maximum: 1000
        reorder_point:
          type: number
          description: Reorder point threshold
          example: 10
          minimum: 0
        primary_supplier_id:
          type: object
          description: Primary supplier ID
          format: uuid
          nullable: true
        supplier_sku:
          type: object
          description: Supplier SKU
          maxLength: 50
          nullable: true
        barcode:
          type: object
          description: Product barcode
          maxLength: 100
          nullable: true
          example: '0641628607549'
        unit:
          type: object
          description: Unit of measure
          maxLength: 50
          nullable: true
          example: units
        is_active:
          type: boolean
          description: Whether the product is active
        is_perishable:
          type: boolean
          description: Whether the product is perishable
        notes:
          type: object
          description: Additional notes
          maxLength: 500
          nullable: true
      required:
        - sku
        - name
        - category_id
        - reorder_point
        - is_active
        - is_perishable
    BulkCreateProductsDto:
      type: object
      properties:
        products:
          description: Array of products to create
          minItems: 1
          maxItems: 100
          type: array
          items:
            $ref: '#/components/schemas/CreateProductDto'
      required:
        - products
...
    BulkDeleteDto:
      type: object
      properties:
        ids:
          description: Array of product IDs to delete
          minItems: 1
          maxItems: 100
          type: array
          items:
            type: string
            format: uuid
        permanent:
          type: boolean
          description: Permanently delete (hard delete) instead of soft delete
          default: false
      required:
        - ids
...